[S12P21A701-328] feature: Create updateCardCompany method

+++ backend/src/main/java/bank/donghang/donghang_api/cardcompany/application/CardCompanyService.java
@@ -1,10 +1,11 @@
 package bank.donghang.donghang_api.cardcompany.application;
 
 import bank.donghang.donghang_api.cardcompany.domain.CardCompany;
-import bank.donghang.donghang_api.cardcompany.domain.repository.CardCompanyJpaRepository;
 import bank.donghang.donghang_api.cardcompany.domain.repository.CardCompanyRepository;
-import bank.donghang.donghang_api.cardcompany.dto.request.CardCompanyCreateRequest;
+import bank.donghang.donghang_api.cardcompany.dto.request.CardCompanyRequest;
 import bank.donghang.donghang_api.cardcompany.dto.response.CardCompanySummaryResponse;
+import bank.donghang.donghang_api.common.exception.BadRequestException;
+import bank.donghang.donghang_api.common.exception.ErrorCode;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,7 +19,7 @@ public class CardCompanyService {
     private final CardCompanyRepository cardCompanyRepository;
 
     @Transactional
-    public Long createCardCompany(CardCompanyCreateRequest request) {
+    public Long createCardCompany(CardCompanyRequest request) {
 
         CardCompany cardCompany = CardCompany.createCardCompany(
                 request.name(),
@@ -36,7 +37,29 @@ public class CardCompanyService {
     }
 
     @Transactional
+    public void updateCardCompany(
+            Long id,
+            CardCompanyRequest cardCompanyUpdateRequest
+    ){
+        if (!cardCompanyRepository.existsCardCompany(id)){
+            throw new BadRequestException(ErrorCode.CARD_COMPANY_NOT_FOUND);
+        }
+
+        CardCompany cardCompany = cardCompanyRepository.findById(id)
+                .orElseThrow(() -> new BadRequestException(ErrorCode.CARD_COMPANY_NOT_FOUND));
+
+        cardCompany.updateCardCompany(
+                cardCompanyUpdateRequest.name(),
+                cardCompanyUpdateRequest.logoUrl()
+        );
+    }
+
+    @Transactional
     public void deleteCardCompany(Long id) {
+
+        if (cardCompanyRepository.existsCardCompany(id)){
+            throw new BadRequestException(ErrorCode.CARD_COMPANY_NOT_FOUND);
+        }
 
         cardCompanyRepository.deleteCardCompany(id);
     }
